requirements.txt
numpy==1.20.0
absl-py==0.13.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
alibi==0.5.8
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
autograd==1.3
bcrypt==3.2.0
blessings==1.7
blis==0.7.4
boto3==1.18.1
botocore==1.21.1
Bottleneck==1.3.2
cached-property==1.5.2
cachetools==4.2.2
catalogue==2.0.4
catboost==0.24.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==7.1.1
cloudpickle==1.6.0
cmdstanpy==0.9.5
colorama==0.4.4
ConfigSpace==0.4.14
contextvars==2.4
convertdate==2.3.2
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
cython==0.29.24
dask==2021.7.0
decorator==4.4.2
dill==0.3.3
distributed==2021.7.0
ephem==4.0.0.2
filelock==3.0.12
flake8==3.9.2
fsspec==2021.7.0
future==0.18.2
gast==0.3.3
google-api-core==1.31.0
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
graphviz==0.8.4
greenlet==1.1.0
grpcio==1.38.1
h5py==2.10.0
HeapDict==1.0.1
hijri-converter==2.1.3
hiredis==2.0.0
holidays==0.11.1
# holoclean==1.0
hyperopt==0.2.5
idna==3.2
imageio==2.9.0
immutables==0.15
importlib-metadata==4.6.1
iniconfig==1.1.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
liac-arff==2.5.0
lightgbm==3.2.1
llvmlite==0.36.0
locket==0.2.1
LunarCalendar==0.0.9
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
minio==7.1.0
msgpack==1.0.2
multidict==5.1.0
multimethod==1.4
murmurhash==1.0.5
mxnet==1.7.0.post1
networkx==2.5.1
numba==0.53.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
openml==0.12.2
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.4
paramiko==2.7.2
partd==1.2.0
pathy==0.6.0
Pillow==8.1.0
plotly==5.1.0
pluggy==0.13.1
portalocker==2.0.0
preshed==3.0.5
prometheus-client==0.11.0
protobuf==3.17.3
psutil==5.7.0
py==1.10.0
py-spy==0.3.7
pyaml==20.4.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
PyMeeus==0.5.11
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
python-dateutil==2.8.2
python-rapidjson==1.4
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
ray[tune]==1.11.0
ray[default]==1.11.0
redis==3.5.3
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
sacrebleu==1.5.1
sacremoses==0.0.45
scikit-image==0.18.2
scikit-learn==1.0.2
scikit-optimize==0.8.1
scipy==1.6.3
sentencepiece==0.1.95
setuptools-git==1.2
shap==0.39.0
six==1.16.0
slicer==0.0.7
smart-open==5.1.0
sortedcontainers==2.4.0
spacy==3.1.0
spacy-legacy==3.0.8
spacy-lookups-data==1.0.2
SQLAlchemy==1.4.21
srsly==2.4.1
tangled-up-in-unicode==0.1.0
tblib==1.7.0
tenacity==8.0.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
thinc==8.0.7
threadpoolctl==2.2.0
tifffile==2021.7.2
tokenizers==0.9.4
toml==0.10.2
toolz==0.11.1
torch==1.7.1
tornado==6.1
tqdm==4.61.2
typer==0.3.2
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
visions==0.7.1
wasabi==0.8.2
Werkzeug==2.0.1
wrapt==1.12.1
xgboost==1.3.3
xmltodict==0.12.0
yacs==0.1.8
yarl==1.6.3
zict==2.0.0
zipp==3.5.0
pytest-timeout==1.4.2
pandasql==0.7.3
fastai==2.3.1
pystan==2.19.1.1
pytest-parallel==0.1.1
river==0.8.0
nvidia-ml-py==11.470.66
rpy2==2.9.6b0

./third_parties/autogluon/core
./third_parties/autogluon/features
./third_parties/autogluon/tabular
./third_parties/autogluon/mxnet
./third_parties/autogluon/text
autogluon-contrib-nlp==0.0.1b20210201

dowhy==0.6
econml==0.11.1
